[CalendarDay 구현]/ 리팩토링 및 구현
