+++ src/App.tsx
@@ -4,11 +4,10 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import './App.css';
 import MainContainer from './components/templetes/MainContainer';
 
-import Mypage from './components/pages/Mypage';
-import Signup from './components/pages/Signup';
-import Login from './components/pages/Login';
-import CalendarDay from './components/pages/CalendarDay';
-import Date from './components/pages/Date';
+import Mypage from './components/templetes/Mypage';
+import Signup from './components/templetes/Signup';
+import Login from './components/templetes/Login';
+import CalendarDay from './components/templetes/CalendarDay';
 import 'bootstrap/dist/css/bootstrap.min.css';
 //css를 불러와야 bootstrap이 적용됨.
 
@@ -17,13 +16,10 @@ function App() {
     <div className="App">
       <Router>
         <Switch>
-          {/* <Route path="/mypage" component={MypageContainer} />
-            <Route path="/login" component={LoginContainer} /> */}
           <Route path="/signup" component={Signup} />
           <Route path="/mypage" component={Mypage} />
           <Route path="/login" component={Login} />
           <Route path="/calendar/day" component={CalendarDay} />
-          <Route path="/calendar/d" component={Date} />
           <Route exact path="/" component={MainContainer} />
         </Switch>
       </Router>
