+++ src/components/UI/Oraganisms/Reviews.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
 export default function Reviews() {
-  return <div></div>;
+  return <div>Here is Review</div>;
 }
