--- src/components/UI/Oraganisms/Todo.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Container, Row } from 'react-bootstrap';
-
-export default function Todo() {
-  return (
-    <Row className="m-1" style={{ border: '1px solid black' }}>
-      Todo
-    </Row>
-  );
-}
