--- src/components/pages/Login.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import LoginContainer from '../templetes/LoginContainer';
-
-export default function Login() {
-  return (
-    <div>
-      <LoginContainer />
-    </div>
-  );
-}
