--- src/components/pages/Signup.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import SignupContainer from '../templetes/SignupContainer';
-
-export default function Signup() {
-  return (
-    <div>
-      <SignupContainer />
-    </div>
-  );
-}
