--- src/components/pages/Todos.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Todos() {
-  return <div></div>;
-}
